agents/agent_service/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
redis
pydantic
numpy
pinecone